Add frontend stub for requesting flair list
Without allow CORS, or giving the frontend the token, we will probably
need to change this so that the server does this for the user.

We may also need to implement some kind of caching for this request, or
some way to limit the rate at which cronnit makes these requests.